Fix a bug in keyBy that didn't take into account multiple indices returning the same key.
When two indices were pointing to the same key and one of them has changed to be a different key, keyBy would have deleted it from the target object, not taking into account that that key is still valid because another key points to it.

Changed the keyBy cache to hold two-way maps (key->index and index->key) so that we can keep track of this.